Move DossierActivateView in a separate module.
